Реализует дополнительное условие о наличии поля address
