+++ app/Classes/FeedbackForm.php
@@ -3,24 +3,10 @@
 namespace App\Classes;
 
 use http\Client\Request;
+use Illuminate\Support\Arr;
 
 class FeedbackForm
 {
-    public static function getInputs (array $formFields): string
-    {
-        if (isset($formFields['profession']) && CheckRules::isStudent($formFields['profession'])) {
-            return 'student';
-        }
 
-        if (isset($formFields['region']) && CheckRules::isCenter($formFields['region'])) {
-            return 'center';
-        }
-
-        if (isset($formFields['product']) && CheckRules::isPromo($formFields['product'])) {
-            return 'promo';
-        }
-
-        return 'all';
-    }
 
 }
