+++ app/Http/Controllers/FeedbackFormController.php
@@ -6,8 +6,8 @@ use App\Classes\CheckRules;
 use App\Classes\FeedbackForm;
 use App\Mail\FormReceived;
 use Illuminate\Http\Response;
+use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\Mail;
-use Illuminate\Support\Facades\Validator;
 use Illuminate\Validation\ValidationException;
 
 class FeedbackFormController extends Controller
@@ -19,9 +19,9 @@ class FeedbackFormController extends Controller
     public function store(): Response
     {
         $data = request()->all();
-        $value = FeedbackForm::getInputs($data);
+        $value = CheckRules::getInputs($data);
 
-        $validator = Validator::make($data, CheckRules::getValidationRules($value));
+        $validator = validator($data, CheckRules::getValidationRules($value, request()));
 
         if ($validator->fails()) {
             return response($validator->errors(), 400);
@@ -29,11 +29,17 @@ class FeedbackFormController extends Controller
 
         $validated = $validator->validate();
 
-//        Mail::to($value . '@example.com')->send(
-//            new FormReceived($validated)
-//        );
 
-        return response($validated,200)
+        CheckRules::hasAddress(request())
+            ? Mail::to($validated['address'])->send(
+                new FormReceived(Arr::except($validated, 'address')))
+            : Mail::to($value . '@example.com')->send(
+                  new FormReceived($validated));
+
+        return CheckRules::hasAddress(request())
+            ? response(Arr::except($validated, 'address'),200)
+                        ->header('Content-Type', 'text/plain')
+            : response($validated,200)
                         ->header('Content-Type', 'text/plain');
     }
 }
